+++ recipes/linux.rb
@@ -1,11 +1,14 @@
-#
+# encoding: utf-8
 # Cookbook Name:: chef-client-upgrade
 # Recipe:: linux
 #
-# Copyright 2013, Nordstrom, Inc.
+# Copyright 2013, 2014 Nordstrom, Inc.
 #
 # All rights reserved - Do Not Redistribute
+#
+# TODO: Find current version, if higher leave it alone.
 
-package node['chef-client-upgrade']['pkg_name'] do
-  action :upgrade
+yum_package node[:chef_client_upgrade][:pkg_name] do
+  action :install
+  version node[:chef_client_upgrade][:chef_pkg_version]
 end