+++ recipes/windows.rb
@@ -1,4 +1,4 @@
-#
+# encoding: utf-8
 # Cookbook Name:: chef-client-upgrade
 # Recipe:: windows
 #
@@ -7,13 +7,13 @@
 # All rights reserved - Do Not Redistribute
 
 nexus        = data_bag_item('shared', 'nexus')
-version      = node["chef-client-upgrade"]["version"]
-package_name = node["chef-client-upgrade"]["display_name"]
-installer    = "chef-client-#{version}.msi"
-url          = node["chef-client-upgrade"]["url"]
-source_url   = "#{nexus["url"]}/#{url}/#{version}/#{installer}"
+version      = node[:chef_client_upgrade][:version]
+package_name = node[:chef_client_upgrade][:display_name]
+installer    = "chef_client_#{version}.msi"
+url          = node[:chef_client_upgrade][:url]
+source_url   = "#{nexus['url']}/#{url}/#{version}/#{installer}"
 
 windows_package package_name do
   source source_url
-  checksum node["chef-client-upgrade"]["msi_checksum"]
+  checksum node[:chef_client_upgrade][:msi_checksum]
 end